WIP: Case of many_many and has_many fix, separation of selection handler
