+++ code/FileAttachmentField.php
@@ -967,7 +967,7 @@ class FileAttachmentField extends FileField {
         }
 
         if($record = $this->getRecord()) {
-            return ($record->many_many($this->getName()) || $record->has_many($this->getName()));
+            return ($record->manyMany($this->getName()) || $record->hasMany($this->getName()));
         }
 
         return false;
@@ -1200,7 +1200,7 @@ class FileAttachmentField extends FileField {
             if (($record = $this->form->getRecord()) && ($record instanceof DataObject)) {
                 $this->record = $record;
             }
-            elseif (($controller = $this->form->Controller())
+            elseif (($controller = $this->form->getController())
                 && $controller->hasMethod('data')
                 && ($record = $controller->data())
                 && ($record instanceof DataObject)
@@ -1334,76 +1334,4 @@ class FileAttachmentField extends FileField {
     }
 }
 
-class FileAttachmentField_SelectHandler extends UploadField_SelectHandler {
 
-    private static $allowed_actions = array (
-        'filesbyid',
-    );
-
-    /**
-     * @param $folderID The ID of the folder to display.
-     * @return FormField
-     */
-    protected function getListField($folderID) {
-        // Generate the folder selection field.
-        $folderField = new TreeDropdownField('ParentID', _t('HtmlEditorField.FOLDER', Folder::class), Folder::class);
-        $folderField->setValue($folderID);
-
-        // Generate the file list field.
-        $config = GridFieldConfig::create();
-        $config->addComponent(new GridFieldSortableHeader());
-        $config->addComponent(new GridFieldFilterHeader());
-        $config->addComponent($columns = new GridFieldDataColumns());
-        $columns->setDisplayFields(array(
-            'StripThumbnail' => '',
-            'Name' => 'Name',
-            'Title' => 'Title'
-        ));
-        $config->addComponent(new GridFieldPaginator(8));
-
-        // If relation is to be autoset, we need to make sure we only list compatible objects.
-        $baseClass = $this->parent->getFileClass();
-
-        // Create the data source for the list of files within the current directory.
-        $files = DataList::create($baseClass)->filter('ParentID', $folderID);
-
-        $fileField = new GridField('Files', false, $files, $config);
-        $fileField->setAttribute('data-selectable', true);
-        if($this->parent->IsMultiple()) {
-            $fileField->setAttribute('data-multiselect', true);
-        }
-
-        $selectComposite = new CompositeField(
-            $folderField,
-            $fileField
-        );
-
-        return $selectComposite;
-    }
-
-
-    public function filesbyid(HTTPRequest $r) {
-        $ids = $r->getVar('ids');
-        $files = File::get()->byIDs(explode(',',$ids));
-
-        $validIDs = array();
-        $json = array ();
-        foreach($files as $file) {
-            $template = new SSViewer('FileAttachmentField_attachments');
-            $html = $template->process(ArrayData::create(array(
-                'File' => $file,
-                'Scope' => $this->parent
-            )));
-
-            $validIDs[$file->ID] = $file->ID;
-            $json[] = array (
-                'id' => $file->ID,
-                'html' => $html->forTemplate()
-            );
-        }
-
-        $this->parent->addValidFileIDs($validIDs);
-        return Convert::array2json($json);
-    }
-
-}
